==> src/DataExploration/0_Install_Packages.R <==
#  List of required packages
list_of_packages <- c('googledrive', 'readr', 'data.table', 'dplyr',
                      'grid', 'gridExtra', 'ggplot2', 'tidyr')

# Install any packages that are not already installed
new_packages <- list_of_packages[!(list_of_packages %in% installed.packages()[,"Package"])]
if (length(new_packages)) {
  install.packages(new_packages)
}
